Move Lexer & Reader inside the Compiler module
