+++ src/php/Command/Shared/NamespaceExtractor.php
@@ -7,8 +7,8 @@ namespace Phel\Command\Shared;
 use Phel\Exceptions\ReaderException;
 use Phel\Lang\Symbol;
 use Phel\Lang\Tuple;
-use Phel\LexerInterface;
-use Phel\ReaderInterface;
+use Phel\Compiler\LexerInterface;
+use Phel\Compiler\ReaderInterface;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
 use RecursiveRegexIterator;
