+++ src/php/Compiler/Token.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Phel;
+namespace Phel\Compiler;
 
 use Phel\Lang\SourceLocation;
 
@@ -33,8 +33,12 @@ final class Token
     private SourceLocation $startLocation;
     private SourceLocation $endLocation;
 
-    public function __construct(int $type, string $code, SourceLocation $startLocation, SourceLocation $endLocation)
-    {
+    public function __construct(
+        int $type,
+        string $code,
+        SourceLocation $startLocation,
+        SourceLocation $endLocation
+    ) {
         $this->type = $type;
         $this->code = $code;
         $this->startLocation = $startLocation;
