+++ src/php/Exceptions/TextExceptionPrinter.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Phel\Exceptions;
 
-use Phel\CodeSnippet;
+use Phel\Compiler\CodeSnippet;
 use Phel\Command\Repl\ColorStyle;
 use Phel\Compiler\Emitter\OutputEmitter\Munge;
 use Phel\Compiler\Emitter\OutputEmitter\SourceMap\SourceMapConsumer;
