+++ tests/php/Unit/ReaderTest.php
@@ -14,8 +14,8 @@ use Phel\Lang\SourceLocation;
 use Phel\Lang\Symbol;
 use Phel\Lang\Table;
 use Phel\Lang\Tuple;
-use Phel\Lexer;
-use Phel\Reader;
+use Phel\Compiler\Lexer;
+use Phel\Compiler\Reader;
 use PHPUnit\Framework\TestCase;
 
 ini_set('xdebug.var_display_max_depth', '10');
